# packages/protocol/src/generated/companion.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "latentPresence companion",
    "version": "1",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "paths": {
...
    "/mcp/tools": {
      "get": {
        "operationId": "mcpTools",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpToolsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Every failure, one shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/call": {
      "post": {
        "operationId": "mcpCall",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpCallResponse"
                }
              }
            }
          },
          "default": {
            "description": "Every failure, one shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpCallRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompanionError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "not_found",
                  "conflict",
                  "unavailable",
                  "database",
                  "upstream",
                  "internal"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "anyOf": [
                  {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {
                      "$ref": "#/$defs/__schema0"
                    }
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
...
      "McpToolsResponse": {
        "type": "object",
        "properties": {
          "servers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1
                },
                "label": {
                  "type": "string",
                  "minLength": 1
                },
                "kind": {
                  "type": "string",
                  "enum": [
                    "stdio",
                    "http",
                    "files"
                  ]
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "starting",
                    "ready",
                    "failed"
                  ]
                },
                "detail": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "instructions": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "id",
                "label",
                "kind",
                "state",
                "detail",
                "instructions"
              ],
              "additionalProperties": false
            }
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "serverId": {
                  "type": "string",
                  "minLength": 1
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "description": {
                  "type": "string"
                },
                "inputSchema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {
                    "$ref": "#/$defs/__schema0"
                  }
                },
                "annotations": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "readOnlyHint": {
                          "type": "boolean"
                        },
                        "destructiveHint": {
                          "type": "boolean"
                        },
                        "idempotentHint": {
                          "type": "boolean"
                        },
                        "openWorldHint": {
                          "type": "boolean"
                        }
                      },
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "serverId",
                "name",
                "description",
                "inputSchema",
                "annotations"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "servers",
          "tools"
        ],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
      "McpCallResponse": {
        "type": "object",
        "properties": {
          "isError": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          },
          "structured": {
            "anyOf": [
              {
                "$ref": "#/$defs/__schema0"
              },
              {
                "type": "null"
              }
            ]
          },
          "images": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "isError",
          "text",
          "structured",
          "images"
        ],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
      "McpCallRequest": {
        "type": "object",
        "properties": {
          "callId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "serverId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "arguments": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "$ref": "#/$defs/__schema0"
            }
          }
        },
        "required": [
          "callId",
          "serverId",
          "name",
          "arguments"
        ],
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
...
    }
  }
}